trying to understand words.txt
i tried to implement the words.txt file into my project, but i still
have a lot of trouble understanding it. my project idea worked as it
should have initially, but i decided to try and take it a step further.
the commit before this is the project and how it was initially, while
this commit is broken after my attempt at trying to implement the file.